+++ frontend/appflowy_flutter/lib/workspace/presentation/widgets/date_picker/utils/layout.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class DatePickerSize {
+  static double scale = 1;
+
+  static double get itemHeight => 26 * scale;
+  static double get seperatorHeight => 4 * scale;
+
+  static EdgeInsets get itemOptionInsets => const EdgeInsets.all(4);
+}
