+++ frontend/appflowy_flutter/lib/workspace/presentation/widgets/date_picker/widgets/clear_date_button.dart
@@ -1,13 +1,12 @@
+import 'package:appflowy/workspace/presentation/widgets/date_picker/utils/layout.dart';
 import 'package:flutter/material.dart';
 
 import 'package:appflowy/generated/locale_keys.g.dart';
-import 'package:appflowy/plugins/database_view/grid/presentation/layout/sizes.dart';
 import 'package:appflowy_popover/appflowy_popover.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra_ui/style_widget/button.dart';
 import 'package:flowy_infra_ui/style_widget/text.dart';
 
-@visibleForTesting
 class ClearDateButton extends StatelessWidget {
   const ClearDateButton({
     super.key,
@@ -21,9 +20,9 @@ class ClearDateButton extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 12.0),
       child: SizedBox(
-        height: GridSize.popoverItemHeight,
+        height: DatePickerSize.itemHeight,
         child: FlowyButton(
-          text: FlowyText.medium(LocaleKeys.grid_field_clearDate.tr()),
+          text: FlowyText.medium(LocaleKeys.datePicker_clearDate.tr()),
           onTap: () {
             onClearDate();
             PopoverContainer.of(context).close();
